Test token refresh
And enable hard session timeout which was
disabled by mistake when moved credentials
from cookies to local storage
