+++ deps/rabbitmq_management/selenium/test/oauth/rabbitmq-localhost.config
@@ -3,13 +3,13 @@
     {auth_backends, [rabbit_auth_backend_oauth2]}
   ]},
    {rabbitmq_management, [
-      {login_session_timeout, 150},   %% in minutes
+      {login_session_timeout, 1},   %% in minutes
       {enable_uaa, true},
       {oauth_enabled, true},
       %{oauth_initiated_logon_type, idp_initiated},
       {oauth_client_id, "rabbit_client_code"},
       %{oauth_client_id, "rabbit_idp_user"},
-    %  {oauth_client_secret, "rabbit_client_code"},
+      {oauth_client_secret, " "},
       %{oauth_provider_url, "http://localhost:3000"}
       {oauth_provider_url, "http://localhost:8080"}
   ]},
