+++ deps/rabbitmq_management/selenium/test/oauth/rabbitmq-with-mgt-prefix.config
@@ -8,7 +8,7 @@
       {oauth_enabled, true},
       {path_prefix, "/my-prefix/another-prefix"},
       {oauth_client_id, "rabbit_client_code"},
-      %{oauth_client_secret, "rabbit_client_code"},
+      {oauth_client_secret, "rabbit_client_code"},
       {oauth_provider_url, "http://uaa:8080"}
   ]},
   {rabbitmq_auth_backend_oauth2, [
