+++ deps/rabbitmq_management/selenium/test/oauth/rabbitmq.config
@@ -7,7 +7,7 @@
       {enable_uaa, true},
       {oauth_enabled, true},
       {oauth_client_id, "rabbit_client_code"},
-      %{oauth_client_secret, "rabbit_client_code"},
+      {oauth_client_secret, "rabbit_client_code"},
       {oauth_provider_url, "http://uaa:8080"}
   ]},
   {rabbitmq_auth_backend_oauth2, [
