+++ deps/rabbitmq_management/selenium/test/oauth/with-idp-initiated-logon/happy-login.js
@@ -16,15 +16,16 @@ describe('A user with a JWT token', function () {
     driver = buildDriver()
     overview = new OverviewPage(driver)
     captureScreen = captureScreensFor(driver, __filename)
+    fakePortal = new FakePortalPage(driver)
   })
 
   it('can log in presenting the token to the /login URL via fakeportal', async function () {
-    fakePortal = new FakePortalPage(driver)
     await fakePortal.goToHome("rabbit_idp_user", "rabbit_idp_user")
     if (!await fakePortal.isLoaded()) {
       throw new Error('Failed to load fakePortal')
     }
     await fakePortal.login()
+    await overview.isLoaded()
     assert.equal(await overview.getUser(), 'User rabbit_idp_user')
   })
